Fixed crashes with team IDs above 127

+++ core/src/mindustry/editor/EditorTile.java
@@ -55,7 +55,7 @@ public class EditorTile extends Tile{
 
         op(OpType.block, block.id);
         if(rotation != 0) op(OpType.rotation, (byte)rotation);
-        if(team() != Team.derelict) op(OpType.team, team().id);
+        if(team() != Team.derelict) op(OpType.team, (byte)team().id);
         super.setBlock(type, team, rotation);
     }
 
@@ -67,7 +67,7 @@ public class EditorTile extends Tile{
         }
 
         if(getTeamID() == team.id) return;
-        op(OpType.team, getTeamID());
+        op(OpType.team, (byte)getTeamID());
         super.setTeam(team);
     }
